refactor: give partial ownership to login component
